Rename potassium slip runtime files
